# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Devot API",
        "version": "0.1"
    },
    "paths": {
...
        "/api/expenses": {
            "get": {
                "tags": [
                    "Expenses"
                ],
                "summary": "Retrieves a collection of expenses for the authenticated user",
                "description": "Retrieves a collection of expenses for the authenticated user.",
                "operationId": "9b702aa9314d471a9de3a121e5183c88",
                "responses": {
                    "200": {
                        "description": "Successful response with a collection of expenses",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/ExpenseResource"
                                    }
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized access",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Unauthenticated."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Expenses"
                ],
                "summary": "Creates a new expense for the authenticated user",
                "description": "Creates a new expense for the authenticated user.",
                "operationId": "a764a04e6d047ab02a68f06c706e67f3",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/StoreExpenseRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Successful response with the newly created expense",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ExpenseResource"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized access",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Unauthenticated."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "402": {
                        "description": "Insufficient funds",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Insufficient funds."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation errors"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/expenses/{expense}": {
            "get": {
                "tags": [
                    "Expenses"
                ],
                "summary": "Retrieves a specific expense for the authenticated user",
                "description": "Retrieves a specific expense for the authenticated user.",
                "operationId": "4fb33b7003219889ff425d4d7e4da429",
                "parameters": [
                    {
                        "name": "expense",
                        "in": "path",
                        "description": "ID of the expense to retrieve",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        },
                        "example": "1"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful response with the expense details",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ExpenseResource"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized access",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Unauthenticated."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Expense not found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Expense not found."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "Expenses"
                ],
                "summary": "Updates an existing expense for the authenticated user",
                "description": "Updates an existing expense for the authenticated user.",
                "operationId": "6769ae6dcdd82c067d7d644c5f55ad61",
                "parameters": [
                    {
                        "name": "expense",
                        "in": "path",
                        "description": "ID of the expense to update",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "description": "New expense details to update",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/UpdateExpenseRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Successful response with the updated expense details",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ExpenseResource"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized access",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Unauthenticated."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Expense not found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Expense not found."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation error(s)",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "The given data was invalid."
                                        },
                                        "errors": {
                                            "type": "object",
                                            "additionalProperties": {
                                                "type": "array",
                                                "items": {
                                                    "type": "string",
                                                    "example": "The :attribute field is required."
                                                }
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Expenses"
                ],
                "summary": "Delete expense",
                "description": "Deletes a specific expense.",
                "operationId": "cb05fc84a7b66979ff8fe43b62e784f3",
                "parameters": [
                    {
                        "name": "expense",
                        "in": "path",
                        "description": "ID of the expense to be deleted",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Expense deleted successfully.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "properties": {
                                                "message": {
                                                    "type": "string",
                                                    "example": "Expense deleted"
                                                },
                                                "item": {
                                                    "$ref": "#/components/schemas/ExpenseResource"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized action.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "error": {
                                            "type": "string",
                                            "example": "Unauthorized action."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
            "Category": {
                "title": "Category",
                "description": "Category",
                "properties": {
                    "id": {
                        "title": "id",
                        "description": "Category Id",
                        "type": "integer",
                        "format": "int64",
                        "example": 1
                    },
                    "name": {
                        "title": "name",
                        "description": "Category name",
                        "type": "string",
                        "example": "Car expenses"
                    }
                },
                "type": "object",
                "xml": {
                    "name": "Category"
                }
            },
            "Expense": {
                "title": "Expense",
                "description": "Expense",
                "properties": {
                    "name": {
                        "title": "name",
                        "description": "Expense name",
                        "type": "string",
                        "example": "Car expenses"
                    },
                    "amount": {
                        "title": "amount",
                        "description": "Expence amount",
                        "type": "number",
                        "format": "float",
                        "example": "124.33"
                    },
                    "description": {
                        "title": "description",
                        "description": "Expense description",
                        "type": "string",
                        "example": "Car expenses for March."
                    },
                    "category": {
                        "title": "id",
                        "description": "Expense category",
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/Category"
                        }
                    },
                    "date": {
                        "title": "id",
                        "description": "Expense category",
                        "type": "string",
                        "example": "27.03.2023"
                    }
                },
                "type": "object",
                "xml": {
                    "name": "Expense"
                }
            },
...
            "StoreExpenseRequest": {
                "title": "Store Expense Request",
                "description": "Store Expense request body data",
                "required": [
                    "name",
                    "category_id",
                    "amount"
                ],
                "properties": {
                    "name": {
                        "title": "name",
                        "description": "Category name",
                        "type": "string",
                        "example": "Car expenses"
                    },
                    "category_id": {
                        "title": "Category id",
                        "description": "Category id",
                        "type": "integer",
                        "example": "1"
                    },
                    "amount": {
                        "title": "amount",
                        "description": "Amount",
                        "type": "number",
                        "format": "float",
                        "example": "77.1"
                    },
                    "description": {
                        "title": "description",
                        "description": "Expense description",
                        "type": "string",
                        "example": "Some expense description....."
                    }
                },
                "type": "object"
            },
...
            "UpdateExpenseRequest": {
                "title": "Store Expense Request",
                "description": "Store Expense request body data",
                "properties": {
                    "name": {
                        "title": "name",
                        "description": "Expense name",
                        "type": "string",
                        "example": "Electricity bill"
                    },
                    "category_id": {
                        "title": "Category id",
                        "description": "Category id",
                        "type": "integer",
                        "example": "1"
                    },
                    "amount": {
                        "title": "amount",
                        "description": "Amount",
                        "type": "number",
                        "format": "float",
                        "example": "77.1"
                    },
                    "description": {
                        "title": "description",
                        "description": "Expense description",
                        "type": "string",
                        "example": "Some expense description....."
                    }
                },
                "type": "object"
            },
...
            "ExpenseResource": {
                "title": "ExpenseResource",
                "description": "Resource category",
                "properties": {
                    "data": {
                        "title": "Data",
                        "description": "Data wrapper",
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/Expense"
                        }
                    }
                },
                "type": "object",
                "xml": {
                    "name": "ExpenseResource"
                }
            }
        }
    }
}